requirements.txt
   Flask==2.0.3
   Flask-Cors==3.0.10
   Flask-Limiter==2.0.0
   Pandas==1.3.4
   NumPy==1.21.2
   Plotly==5.1.0
   Streamlit==1.0.0
   Requests==2.26.0
   python-dotenv==0.19.1
   Psutil==5.8.0
   ccxt  # 使用最新版本
   SQLAlchemy==1.4.22
   scipy>=1.7.0  # 用于期权计算
   APScheduler>=3.10.0  # 用于定时任务